requirements.txt
fastapi
uvicorn
python-multipart
pillow
aiofiles
tqdm
matplotlib
seaborn
pandas
requests
opencv-python-headless==4.6.0.66
gdown
